Tool path frag shader added/img data refactor
I had to refactor the way the script stores the
image data. I was storing the images retrieved from
ModernGL as numpy arrays, but this led to translation
problems if I wanted to feed that image into a different
buffer from the CPU. I decided to store all images as
byte data and only convert once the images need to be
saved.

Also, "cutting" the image is now possible in the computeWorker
class using a simple frag shader. The work to implement this
revealed the errors that I had to fix in the refactor stated
above.